move netcdf_output_writer.jl to extension
